25 Personality items representing 5 factors

25 personality self report items taken from the International Personality Item Pool (ipip.ori.org) were included as part of the Synthetic Aperture Personality Assessment (SAPA) web based personality assessment project.

bfi

Format

The data from 2800 subjects.

A1:

Am indifferent to the feelings of others. (q_146)

A2:

Inquire about others' well-being. (q_1162)

A3:

Know how to comfort others. (q_1206)

A4:

Love children. (q_1364)

A5:

Make people feel at ease. (q_1419)

C1:

Am exacting in my work. (q_124)

C2:

Continue until everything is perfect. (q_530)

C3:

Do things according to a plan. (q_619)

C4:

Do things in a half-way manner. (q_626)

C5:

Waste my time. (q_1949)

E1:

Don't talk a lot. (q_712)

E2:

Find it difficult to approach others. (q_901)

E3:

Know how to captivate people. (q_1205)

E4:

Make friends easily. (q_1410)

E5:

Take charge. (q_1768)

N1:

Get angry easily. (q_952)

N2:

Get irritated easily. (q_974)

N3:

Have frequent mood swings. (q_1099)

N4:

Often feel blue. (q_1479)

N5:

Panic easily. (q_1505)

O1:

Am full of ideas. (q_128)

O2:

Avoid difficult reading material.(q_316)

O3:

Carry the conversation to a higher level. (q_492)

O4:

Spend time reflecting on things. (q_1738)

O5:

Will not probe deeply into a subject. (q_1964)

Details

This is the same as the bfi in the psych package, except with reverse coded items reversed as they would need to be for scoring purposes. I also drop the demographic variables.

See also

Examples

library(visibly)
data(bfi)
summary(bfi)

#>        A1               A2              A3              A4            A5      
#>  Min.   :-6.000   Min.   :1.000   Min.   :1.000   Min.   :1.0   Min.   :1.00  
#>  1st Qu.:-3.000   1st Qu.:4.000   1st Qu.:4.000   1st Qu.:4.0   1st Qu.:4.00  
#>  Median :-2.000   Median :5.000   Median :5.000   Median :5.0   Median :5.00  
#>  Mean   :-2.413   Mean   :4.802   Mean   :4.604   Mean   :4.7   Mean   :4.56  
#>  3rd Qu.:-1.000   3rd Qu.:6.000   3rd Qu.:6.000   3rd Qu.:6.0   3rd Qu.:5.00  
#>  Max.   :-1.000   Max.   :6.000   Max.   :6.000   Max.   :6.0   Max.   :6.00  
#>  NA's   :16       NA's   :27      NA's   :26      NA's   :19    NA's   :16    
#>        C1              C2             C3              C4        
#>  Min.   :1.000   Min.   :1.00   Min.   :1.000   Min.   :-6.000  
#>  1st Qu.:4.000   1st Qu.:4.00   1st Qu.:4.000   1st Qu.:-4.000  
#>  Median :5.000   Median :5.00   Median :5.000   Median :-2.000  
#>  Mean   :4.502   Mean   :4.37   Mean   :4.304   Mean   :-2.553  
#>  3rd Qu.:5.000   3rd Qu.:5.00   3rd Qu.:5.000   3rd Qu.:-1.000  
#>  Max.   :6.000   Max.   :6.00   Max.   :6.000   Max.   :-1.000  
#>  NA's   :21      NA's   :24     NA's   :20      NA's   :26      
#>        C5               E1               E2               E3       
#>  Min.   :-6.000   Min.   :-6.000   Min.   :-6.000   Min.   :1.000  
#>  1st Qu.:-5.000   1st Qu.:-4.000   1st Qu.:-4.000   1st Qu.:3.000  
#>  Median :-3.000   Median :-3.000   Median :-3.000   Median :4.000  
#>  Mean   :-3.297   Mean   :-2.974   Mean   :-3.142   Mean   :4.001  
#>  3rd Qu.:-2.000   3rd Qu.:-2.000   3rd Qu.:-2.000   3rd Qu.:5.000  
#>  Max.   :-1.000   Max.   :-1.000   Max.   :-1.000   Max.   :6.000  
#>  NA's   :16       NA's   :23       NA's   :16       NA's   :25     
#>        E4              E5              N1              N2       
#>  Min.   :1.000   Min.   :1.000   Min.   :1.000   Min.   :1.000  
#>  1st Qu.:4.000   1st Qu.:4.000   1st Qu.:2.000   1st Qu.:2.000  
#>  Median :5.000   Median :5.000   Median :3.000   Median :4.000  
#>  Mean   :4.422   Mean   :4.416   Mean   :2.929   Mean   :3.508  
#>  3rd Qu.:6.000   3rd Qu.:5.000   3rd Qu.:4.000   3rd Qu.:5.000  
#>  Max.   :6.000   Max.   :6.000   Max.   :6.000   Max.   :6.000  
#>  NA's   :9       NA's   :21      NA's   :22      NA's   :21     
#>        N3              N4              N5             O1       
#>  Min.   :1.000   Min.   :1.000   Min.   :1.00   Min.   :1.000  
#>  1st Qu.:2.000   1st Qu.:2.000   1st Qu.:2.00   1st Qu.:4.000  
#>  Median :3.000   Median :3.000   Median :3.00   Median :5.000  
#>  Mean   :3.217   Mean   :3.186   Mean   :2.97   Mean   :4.816  
#>  3rd Qu.:4.000   3rd Qu.:4.000   3rd Qu.:4.00   3rd Qu.:6.000  
#>  Max.   :6.000   Max.   :6.000   Max.   :6.00   Max.   :6.000  
#>  NA's   :11      NA's   :36      NA's   :29     NA's   :22     
#>        O2               O3              O4              O5       
#>  Min.   :-6.000   Min.   :1.000   Min.   :1.000   Min.   :-6.00  
#>  1st Qu.:-4.000   1st Qu.:4.000   1st Qu.:4.000   1st Qu.:-3.00  
#>  Median :-2.000   Median :5.000   Median :5.000   Median :-2.00  
#>  Mean   :-2.713   Mean   :4.438   Mean   :4.892   Mean   :-2.49  
#>  3rd Qu.:-1.000   3rd Qu.:5.000   3rd Qu.:6.000   3rd Qu.:-1.00  
#>  Max.   :-1.000   Max.   :6.000   Max.   :6.000   Max.   :-1.00  
#>                   NA's   :28      NA's   :14      NA's   :20