Plot fixed or random effects coefficients for brmsfit objects.

Plot fixed or random effects coefficients for brmsfit objects.

# S3 method for brmsfit
plot_coefficients(
  model,
  order = "decreasing",
  sd_multi = 2,
  keep_intercept = FALSE,
  palette = "bilbao",
  ref_line = 0,
  trans = NULL,
  plot = TRUE,
  ranef = FALSE,
  which_ranef = NULL,
  ...
)

Arguments

model
The model that is the point of this function. For example, lm, glm, gam, lme4, brms.

order
The order of the plots- "increasing", "decreasing", or a numeric vector giving the order. The default is NULL, i.e. the default ordering. Not applied to random effects.

sd_multi
The multiplier that determines the width of the interval. Default is 2.

keep_intercept
Default is FALSE. Intercepts are typically on a very different scale than covariate effects.

palette
A scico palette. Default is 'bilbao'.

ref_line
A reference line. Default is zero.

trans
A transformation function to be applied to the coefficients (e.g. exponentiation).

plot
Default is TRUE, but sometimes you just want the data.

ranef
If applicable, whether to plot random effects instead of fixed effects.

which_ranef
If plotting random effects, which one to plot.

...
Other arguments applied for specific methods.

Value

A ggplot of the coefficients and their interval estimates. Or the data that would be used to create the plot.

See also

Examples

# placeholder